TM-135: fix runner on Windows
The problem was that picocli was interpreting everything starting with
``\`` as an escape sequence. More specifically, it was using StreamTokenizer
setup with quote characters set to both ``'`` and ``"``. Everything
inside quotes will be interpreted as an escape sequence, and must be
escaped.

Also got rid of the "toWindowsFriendly()" function which was only
a workaround to this issue.