TM-207 Test doesn't run - java.lang.OutOfMemoryError: Java heap space
The problem was that we didn't close the WebSocket connection after the
test finished executing. The WebSocket-connection-related classes have
an indirect reference to a buffer which we set to approximately 300MB.

Since we kept the connection open, and we opened a new WebSocket
connection every time we ran some tests, after a few runs, we exhausted
the memory.