TM-414 [backend] test: Can't delete a substep
I could reproduce the problem when when filename escaping rules changed
since the file was created. In my case, we added the rule that filenames
are not allowed to end in space (creates issues on Windows).
When updating such a file, the "old file" was not found, and a new file
was created instead (having the updated content), leaving the original
file untouched.